requirements.txt

python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-telegram-bot
playwright
beautifulsoup4 
lxml

# for recaptcha
urllib3
SpeechRecognition
pydub
fake-useragent